classquiz_backend/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.5.4
openai==1.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
Pillow==10.3.0
httpx==0.27.0
tenacity==8.3.0
structlog==24.1.0
